FindBug

+++ fods-core/src/main/java/org/jsoftware/fods/impl/stats/StatisticsItem.java
@@ -33,7 +33,7 @@ public class StatisticsItem implements Serializable {
 
 
 	public JMXStatistics createJMXStatistics(String myDbName) {
-		double avrSelectQyeryMS = (executedSelectQueries == 0) ? 0 : executedSelectTime / executedSelectQueries;
+		double avrSelectQyeryMS = (executedSelectQueries == 0) ? 0 : (double)executedSelectTime / (double)executedSelectQueries;
 		double avrQyeryMS = (executedQueries == 0) ? 0 : (double) executedTime / (double) executedQueries;
 		return new JMXStatistics(get, release, breakTimes, recoveryTimes, executedQueries, executedSelectQueries, myDbName, avrSelectQyeryMS, avrQyeryMS);
 	}
